requirements.txt
requests
beautifulsoup4
ipykernel
biopython